Log 'panic' when using upstart
We don't have a way to alway log panics from the agent itself. So with
upstart we need to activate the 'log' option for the console. This way
panic logs will end up in /var/log/upstart/datadog-agent.log

By default systemd will redirect stderr to the journal, so we don't have
to specify anything.